backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.3.0"
description = ""
authors = ["Alexander Demure"]

[tool.poetry.dependencies]

python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.5"
structlog = "^20.1.0"
tenacity = "^6.2.0"

httpx = "^0.16.1"

email-validator = "^1.1.1"
jinja2 ="^2.11.2"

fastapi-auth = "^0.1.1"
alchemy-permissions = "^0.0.5"
security-utils = "^0.0.1"


fastapi-sqlalchemy = "^0.1.4"
psycopg2 = "^2.8.5"
alembic = "^1.4.2"
databases = "^0.3.2"
asyncpg = "^0.20.1"


[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
